+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changes for lovely.esdb
 unreleased
 ==========
 
+ - added static method `bulk_update` to update multiple docuemnts with one
+   single bulk request
+
 2015/12/11 0.1.0
 ================